suppression et ajout de commentaires + changement des noms des classes car elles n'ont plus la même fonctionnalité
